# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateExpenseResponse:
    properties:
      ai_audit_notes:
        example: No policy violations detected.
        type: string
      ai_fraud_flag:
        example: false
        type: boolean
      expense_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  handlers.CreateJournalEntryResponse:
    properties:
      entry_number:
        example: JE-20260102-0042
        type: string
      journal_entry_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  handlers.OCRResponse:
    properties:
      invoice_number:
        example: INV-4521
        type: string
      line_items:
        items:
          $ref: '#/definitions/services.OCRLineItem'
        type: array
      tax_amount:
        example: 45.6
        type: number
      total_amount:
        example: 570
        type: number
      vendor_name:
        example: Acme Supplies Inc.
        type: string
    type: object
...
  handlers.calculateTaxRequest:
    properties:
      country_code:
        type: string
      subtotal:
        type: number
    type: object
...
  handlers.createAssetRequest:
    properties:
      asset_name:
        type: string
      purchase_cost:
        type: number
      purchase_date:
        type: string
      salvage_value:
        type: number
      useful_life_years:
        type: integer
    type: object
...
  handlers.createExpenseRequest:
    properties:
      amount:
        type: number
      category:
        type: string
      receipt_file:
        type: string
    type: object
  handlers.createJournalEntryRequest:
    properties:
      entry_date:
        type: string
      items:
        items:
          $ref: '#/definitions/handlers.journalItemRequest'
        type: array
      memo:
        type: string
    type: object
...
  handlers.createTaxRateRequest:
    properties:
      country_code:
        type: string
      tax_name:
        type: string
      tax_rate:
        type: number
    type: object
...
  handlers.journalItemRequest:
    properties:
      account_id:
        type: string
      credit:
        type: number
      debit:
        type: number
    type: object
...
  services.OCRLineItem:
    properties:
      description:
        type: string
      quantity:
        type: integer
      total:
        type: number
      unit_price:
        type: number
    type: object
...
  utils.Envelope:
    additionalProperties: {}
    type: object
host: localhost:8080
info:
  contact:
    email: support@strata.dev
    name: Strata Support
  description: ERP-CRM Hybrid API — direct competitor to Odoo.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Strata API
  version: 0.1.0
paths:
...
  /api/v1/accounting/assets:
    post:
      consumes:
      - application/json
      description: Registers a new fixed asset for depreciation tracking. Requires
        `accounting.assets.write` permission.
      operationId: createAsset
      parameters:
      - description: Asset payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createAssetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Register fixed asset
      tags:
      - Accounting
  /api/v1/accounting/assets/{asset_id}/depreciation:
    get:
      description: Calculates straight-line depreciation for a fixed asset over a
        given date range. Returns annual depreciation, accumulated depreciation, and
        current book value. Requires `accounting.assets.read` permission.
      operationId: getDepreciation
      parameters:
      - description: Asset ID
        in: path
        name: asset_id
        required: true
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: to_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Calculate asset depreciation
      tags:
      - Accounting
...
  /api/v1/accounting/expenses:
    post:
      consumes:
      - application/json
      description: Creates an expense submission and runs AI fraud detection. Flags
        high-value, policy-violating, or suspicious expenses. Requires `expenses.submit`
        permission.
      operationId: createExpense
      parameters:
      - description: Expense payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createExpenseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateExpenseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Submit expense with AI fraud audit
      tags:
      - Accounting
  /api/v1/accounting/invoices/ocr:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a PDF or image invoice file for AI-powered OCR extraction.
        Returns extracted vendor, line items, tax, and total. Requires `accounting.invoices.write`
        permission.
      operationId: processInvoiceOCR
      parameters:
      - description: Invoice file (PDF or image)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.OCRResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Upload invoice for vision OCR
      tags:
      - Accounting
  /api/v1/accounting/journal-entries:
    post:
      consumes:
      - application/json
      description: Creates a journal entry with balanced debit/credit items. Total
        debits must equal total credits. Requires `accounting.ledger.write` permission.
      operationId: createJournalEntry
      parameters:
      - description: Journal entry payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createJournalEntryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateJournalEntryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Post general ledger entry
      tags:
      - Accounting
  /api/v1/accounting/tax-rates:
    post:
      consumes:
      - application/json
      description: Creates a new tax rate for a specific country (e.g., VAT, GST).
        Requires `accounting.tax.manage` permission.
      operationId: createTaxRate
      parameters:
      - description: Tax rate payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createTaxRateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Create tax rate
      tags:
      - Accounting
  /api/v1/accounting/tax/calculate:
    post:
      consumes:
      - application/json
      description: Computes tax for a given country code and subtotal amount using
        the organization's active tax rates. Requires `accounting.tax.read` permission.
      operationId: calculateTax
      parameters:
      - description: Tax calculation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.calculateTaxRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Calculate tax
      tags:
      - Accounting
...
securityDefinitions:
  ApiKeyAuth:
    description: API key for programmatic access (e.g. fleet telematics ingestion)
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'Enter your Bearer token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...